修改源码

+++ handler/src/test/java/io/netty/handler/stream/ChunkedWriteHandlerTest.java
@@ -213,7 +213,7 @@ public class ChunkedWriteHandlerTest {
         ch.checkException();
         assertTrue(ch.finish());
 
-        assertEquals(0, ch.readOutbound());
+        assertEquals(0L, (long) ch.readOutbound());
         assertNull(ch.readOutbound());
     }
